site: extract trail rendering primitives from AnimatedTrails
